feat(coqui_english_huge_vocab): add STT coqui_english_huge_vocab model (#423)
* fix(submodule): fix audio references in submodules.py

* feat(CoquiEngineHelper): add CoquiEngineHelper for STT

* fix(CoquiEngineHelper): fix model_uri reference in CoquiEngineHelper

* fix(CoquiEngineHelper): fix init in CoquiEngineHelper

* feat(coqui_english_huge_vocab): aadd STT coqui_english_huge_vocab model

* fix(env.yaml): fix missing import for stt in env.yaml

* test(audio): add missing unit test ffor audio modality

* fix(CoquiEngineHelper): wrong scorer reference in CoquiEngineHelper

* fix(download_model): fix lots of issues in download model

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update CoquiEngineHelper.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix(coqui_english_huge_vocab): remove concurrent execution

+++ src/api_utils/gladia_api_utils/file_management.py
@@ -90,21 +90,22 @@ def write_url_content_to_file(file_full_path: Path, url) -> bool:
     return write_to_file(file_full_path, data)
 
 
-def write_to_file(file_full_path, data) -> bool:
+def write_to_file(file_full_path, data, overwrite=False) -> bool:
     if isinstance(file_full_path, str):
         file_full_path = Path(file_full_path)
 
     os.makedirs(os.path.dirname(file_full_path), exist_ok=True)
 
-    if isinstance(data, io.BytesIO):
-        with open(file_full_path, "wb") as handler:
-            data = data.read()
-        handler.write(data)
-    elif isinstance(data, Image.Image):
-        data.save(file_full_path)
-    else:
-        with open(file_full_path, "wb") as handler:
+    if not os.path.exists(file_full_path) or overwrite:
+        if isinstance(data, io.BytesIO):
+            with open(file_full_path, "wb") as handler:
+                data = data.read()
             handler.write(data)
+        elif isinstance(data, Image.Image):
+            data.save(file_full_path)
+        else:
+            with open(file_full_path, "wb") as handler:
+                handler.write(data)
 
 
 def generate_random_filename(upload_directory, extension):
